requirements.txt
pyyaml~=6.0
dataset~=1.5.2
python-telegram-bot

requests~=2.28.1